requirements.txt
discord.py==1.5.1
discord-ext-menus==1.1
git+https://github.com/Ext-Creators/discord-ext-events
umongo==3.0.0
motor==2.4.0
pynacl==1.4.0
pytz==2019.3
requests